A few miscellaneous fixes

+++ Nernel/Include/CLR/InternalCalls/System.DateTime.h
@@ -2,4 +2,4 @@
 
 #include <CLR/AppDomain.h>
 
-uint64_t System_DateTime_InternalUtcNow(AppDomain* pAppDomain);
+uint32_t System_DateTime_InternalUtcNow(AppDomain* pAppDomain);
